+++ chirho/robust/ops.py
@@ -6,11 +6,9 @@ from typing import (
     Mapping,
     Optional,
     ParamSpec,
-    Tuple,
     TypeVar,
 )
 
-import pyro
 import torch
 from chirho.observational.ops import Observation
 
@@ -25,18 +23,6 @@ Functional = Callable[[Model[P], Model[P]], Callable[P, S]]
 ParamDict = Mapping[str, torch.Tensor]
 
 
-def make_functional_call(
-    mod: Callable[P, T]
-) -> Tuple[ParamDict, Callable[Concatenate[ParamDict, P], T]]:
-    assert isinstance(mod, torch.nn.Module)
-    param_dict: ParamDict = dict(mod.named_parameters())
-    mod_func: Callable[Concatenate[ParamDict, P], T] = functools.partial(torch.func.functional_call, mod)
-    functionalized_mod_func: Callable[Concatenate[ParamDict, P], T] = torch.func.functionalize(
-        pyro.validation_enabled(False)(mod_func)
-    )
-    return param_dict, functionalized_mod_func
-
-
 def influence_fn(
     model: Model[P],
     guide: Model[P],
@@ -44,7 +30,7 @@ def influence_fn(
     **linearize_kwargs
 ) -> Callable[Concatenate[Point[T], P], S]:
 
-    from chirho.robust.internals import linearize
+    from chirho.robust.internals import linearize, make_functional_call
 
     linearized = linearize(model, guide, **linearize_kwargs)
